+++ resources/views/admin/asset/update.php
@@ -3,27 +3,28 @@
 declare(strict_types=1);
 
 /**
- * @see EntryAssetController::actionUpdate()
+ * @see AssetController::actionUpdate()
  *
  * @var View $this
  * @var Asset $asset
  */
 
 use Hirtz\Cms\Models\Asset;
-use Hirtz\Cms\Modules\Admin\Controllers\EntryAssetController;
+use Hirtz\Cms\Modules\Admin\Controllers\AssetController;
 use Hirtz\Cms\Modules\Admin\Widgets\Forms\AssetActiveForm;
 use Hirtz\Cms\Modules\Admin\Widgets\Navs\AssetActionDropdown;
-use Hirtz\Cms\Modules\Admin\Widgets\Navs\EntryHeader;
-use Hirtz\Cms\Modules\Admin\Widgets\Navs\EntrySubmenu;
+use Hirtz\Cms\Modules\Admin\Widgets\Navs\AssetHeader;
+use Hirtz\Cms\Modules\Admin\Widgets\Navs\AssetSubmenu;
 use Hirtz\Skeleton\Web\View;
 use Hirtz\Skeleton\Widgets\Forms\FormContainer;
 
-echo EntryHeader::make()
-    ->content(AssetActionDropdown::make()->model($asset))
-    ->model($asset->entry);
+echo AssetHeader::make()
+    ->model($asset->parent)
+    ->content(AssetActionDropdown::make()
+        ->model($asset));
 
-echo EntrySubmenu::make()
-    ->model($asset->entry);
+echo AssetSubmenu::make()
+    ->model($asset->parent);
 
 echo FormContainer::make()
     ->title($this->title)