+++ src/Modules/Admin/Data/AssetArrayDataProvider.php
@@ -4,17 +4,14 @@ declare(strict_types=1);
 
 namespace Hirtz\Cms\Modules\Admin\Data;
 
-use Hirtz\Cms\Models\Category;
+use Hirtz\Cms\Models\Asset;
 use Hirtz\Cms\Models\Entry;
-use Hirtz\Cms\Models\Queries\CategoryQuery;
 use Hirtz\Cms\Models\Section;
-use Hirtz\Skeleton\Base\Traits\ModelTrait;
 use yii\data\ArrayDataProvider;
 
 /**
- * @property CategoryQuery $query
- * @property Category[] $models
- * @method Category[] getModels()
+ * @property Asset[] $models
+ * @method Asset[] getModels()
  */
 class AssetArrayDataProvider extends ArrayDataProvider
 {